--- src/main/java/bowling/domain/BowlingGameHitResultIdentification.java
@@ -1,10 +1,0 @@
-package bowling.domain;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface BowlingGameHitResultIdentification {
-
-    boolean identify(List<Integer> hits);
-
-}
