+++ src/main/java/bowling/domain/frame/FinalBowlingGameFrame.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.frame;
+
+import bowling.domain.state.BowlingGameHitState;
 
 public class FinalBowlingGameFrame extends AbstractBowlingGameFrame {
